mico algotop BFS(4)

+++ algorithm/week2/day2/02_minco_algotop_BFS(4).cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 struct HP { int to, hp; };
 vector<HP> adjList[100001];
-int visited[100001];
+long long visited[100001];
 
 
 int main() {
